Move `finalize and report timing` to controller
A new method called `conclude` reports the timing for the resource.
The timing info used is the one attached to the fetch rather than
one attached to a response.

This is done to reduce ambiguities when timing info is attached to
the `response`, as it's unclear what happens to that timing info
when a response is cloned, reused, serialized or stored & restored
from cache.